[REFACTOR] 오늘 날씨 예보 -> 현재기온으로 수정
